MysteryNumberDuel.java
So what the program does is the bot has 10 choices 1 to 10 and you also have to enter numbers 1 to 10, the bot chooses a number from the array of type int with nunbers 1 to 10, then if you and the bot choose the same number you lose. If you choose a number that's different from the bot you win but if thst number you chose is less then the number that the bot chose you lose.